refactor(features): move `profile` folder from `home`
